Ok I dont know if either of these are working, This one might work though XD.  I am commit this untested stuff because some people need the structures for development.
Hi Alex / Noor,

Im going to crack on with Testing today, I am not sure if my file structure is messed up so am am only comitting the new files that I have written.

I hope to test these today!

PS. The Identifier code is literally straight from the android site.... but it is perfect. I dont know what to do about that, do we just refactor it in a cheeky manner or do we keep it as is and declare where it came from?

Ed